county_portal/requirements.txt
alembic==1.16.1
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-Principal==0.4.0
Flask-Security==5.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.2
idna==3.10
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
python-dotenv==1.1.0
SQLAlchemy==2.0.41
tomli==2.2.1
typing_extensions==4.13.2
Werkzeug==3.1.3
WTForms==3.2.1
